Land Recent QUIC changes.
QUIC: support for rotating QUIC server configs.

In the limit, we would want to rotate QUIC server configs in the same fashion
as we rotate TLS SessionTicket keys because the initial client application data
isn't forward secure.

Merge internal change: 47706026

QUIC: fix flakey test.

Once in every 256 test runs, the first byte of the box would actually be 'X'.

Merge internal change: 47653885

Disabling LargePostFEC test until b/9295090 is Fixed.  Remaining flakiness
should be fixed by rch's pending changes.

Merge internal change: 47649448

Merging changes from chromium CL - 16256017

Merge internal change: 47608845

Changing reliable quic stream's Readv and users from int to size_t to avoid
potential size mismatch issues.

Merge internal change: 47475275

R=rch@chromium.org

Review URL: https://chromiumcodereview.appspot.com/16747002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@205531 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d6ce2eb2dfaf06f5effc46c2a4e0b652a983c6b4

+++ net/quic/reliable_quic_stream.cc
@@ -103,7 +103,7 @@ void ReliableQuicStream::Close(QuicRstStreamErrorCode error) {
   }
 }
 
-int ReliableQuicStream::Readv(const struct iovec* iov, size_t iov_len) {
+size_t ReliableQuicStream::Readv(const struct iovec* iov, size_t iov_len) {
   if (headers_decompressed_ && decompressed_headers_.empty()) {
     return sequencer_.Readv(iov, iov_len);
   }
@@ -111,8 +111,8 @@ int ReliableQuicStream::Readv(const struct iovec* iov, size_t iov_len) {
   size_t iov_index = 0;
   while (iov_index < iov_len &&
          decompressed_headers_.length() > bytes_consumed) {
-    int bytes_to_read = min(iov[iov_index].iov_len,
-                            decompressed_headers_.length() - bytes_consumed);
+    size_t bytes_to_read = min(iov[iov_index].iov_len,
+                               decompressed_headers_.length() - bytes_consumed);
     char* iov_ptr = static_cast<char*>(iov[iov_index].iov_base);
     memcpy(iov_ptr,
            decompressed_headers_.data() + bytes_consumed, bytes_to_read);
